Extract out diagnostic sources into their own files
