+++ src/Features/LanguageServer/Protocol/Handler/Diagnostics/WorkspacePullDiagnosticHandler.cs
@@ -28,7 +28,7 @@ using Range = Microsoft.VisualStudio.LanguageServer.Protocol.Range;
 namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics
 {
     [Method(VSInternalMethods.WorkspacePullDiagnosticName)]
-    internal sealed class WorkspacePullDiagnosticHandler : AbstractPullDiagnosticHandler<VSInternalWorkspaceDiagnosticsParams, VSInternalWorkspaceDiagnosticReport, VSInternalWorkspaceDiagnosticReport[]>
+    internal sealed partial class WorkspacePullDiagnosticHandler : AbstractPullDiagnosticHandler<VSInternalWorkspaceDiagnosticsParams, VSInternalWorkspaceDiagnosticReport, VSInternalWorkspaceDiagnosticReport[]>
     {
         public WorkspacePullDiagnosticHandler(IDiagnosticAnalyzerService analyzerService, EditAndContinueDiagnosticUpdateSource editAndContinueDiagnosticUpdateSource, IGlobalOptionService globalOptions)
             : base(analyzerService, editAndContinueDiagnosticUpdateSource, globalOptions)
@@ -67,7 +67,7 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics
 
         protected override ValueTask<ImmutableArray<IDiagnosticSource>> GetOrderedDiagnosticSourcesAsync(RequestContext context, CancellationToken cancellationToken)
         {
-            return GetWorkspacePullDocumentsAsync(context, GlobalOptions, cancellationToken);
+            return GetDiagnosticSourcesAsync(context, GlobalOptions, cancellationToken);
         }
 
         protected override VSInternalWorkspaceDiagnosticReport[]? CreateReturn(BufferedProgress<VSInternalWorkspaceDiagnosticReport> progress)
@@ -75,7 +75,7 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics
             return progress.GetValues();
         }
 
-        internal static async ValueTask<ImmutableArray<IDiagnosticSource>> GetWorkspacePullDocumentsAsync(RequestContext context, IGlobalOptionService globalOptions, CancellationToken cancellationToken)
+        internal static async ValueTask<ImmutableArray<IDiagnosticSource>> GetDiagnosticSourcesAsync(RequestContext context, IGlobalOptionService globalOptions, CancellationToken cancellationToken)
         {
             Contract.ThrowIfNull(context.Solution);
 
@@ -163,64 +163,6 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics
                 if (isFSAOn)
                 {
                     result.Add(new ProjectDiagnosticSource(project));
-                }
-            }
-        }
-
-        private record struct ProjectDiagnosticSource(Project Project) : IDiagnosticSource
-        {
-            public ProjectOrDocumentId GetId() => new(Project.Id);
-
-            public Project GetProject() => Project;
-
-            public Uri GetUri()
-            {
-                Contract.ThrowIfNull(Project.FilePath);
-                return ProtocolConversions.GetUriFromFilePath(Project.FilePath);
-            }
-
-            public async Task<ImmutableArray<DiagnosticData>> GetDiagnosticsAsync(
-                IDiagnosticAnalyzerService diagnosticAnalyzerService,
-                RequestContext context,
-                DiagnosticMode diagnosticMode,
-                CancellationToken cancellationToken)
-            {
-                // Directly use the IDiagnosticAnalyzerService.  This will use the actual snapshots
-                // we're passing in.  If information is already cached for that snapshot, it will be returned.  Otherwise,
-                // it will be computed on demand.  Because it is always accurate as per this snapshot, all spans are correct
-                // and do not need to be adjusted.
-                var projectDiagnostics = await diagnosticAnalyzerService.GetProjectDiagnosticsForIdsAsync(Project.Solution, Project.Id, cancellationToken: cancellationToken).ConfigureAwait(false);
-                return projectDiagnostics;
-            }
-        }
-
-        private record struct WorkspaceDocumentDiagnosticSource(TextDocument Document) : IDiagnosticSource
-        {
-            public ProjectOrDocumentId GetId() => new(Document.Id);
-
-            public Project GetProject() => Document.Project;
-
-            public Uri GetUri() => Document.GetURI();
-
-            public async Task<ImmutableArray<DiagnosticData>> GetDiagnosticsAsync(
-                IDiagnosticAnalyzerService diagnosticAnalyzerService,
-                RequestContext context,
-                DiagnosticMode diagnosticMode,
-                CancellationToken cancellationToken)
-            {
-                if (Document is SourceGeneratedDocument sourceGeneratedDocument)
-                {
-                    // Unfortunately GetDiagnosticsForIdsAsync returns nothing for source generated documents.
-                    var documentDiagnostics = await diagnosticAnalyzerService.GetDiagnosticsForSpanAsync(sourceGeneratedDocument, range: null, cancellationToken: cancellationToken).ConfigureAwait(false);
-                    return documentDiagnostics;
-                }
-                else
-                {
-                    // We call GetDiagnosticsForIdsAsync as we want to ensure we get the full set of diagnostics for this document
-                    // including those reported as a compilation end diagnostic.  These are not included in document pull (uses GetDiagnosticsForSpan) due to cost.
-                    // However we can include them as a part of workspace pull when FSA is on.
-                    var documentDiagnostics = await diagnosticAnalyzerService.GetDiagnosticsForIdsAsync(Document.Project.Solution, Document.Project.Id, Document.Id, cancellationToken: cancellationToken).ConfigureAwait(false);
-                    return documentDiagnostics;
                 }
             }
         }
